amplify/functions/aws-documentation/requirements.txt
awslabs.aws-documentation-mcp-server==0.1.4